[BE] FIX : 본인의 게시글이 아닌 경우에 대한 exception 추가

+++ backend/src/main/java/proj/pet/board/service/BoardServiceImpl.java
@@ -24,8 +24,7 @@ import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
-import static proj.pet.exception.ExceptionStatus.NOT_FOUND_BOARD;
-import static proj.pet.exception.ExceptionStatus.NOT_FOUND_MEMBER;
+import static proj.pet.exception.ExceptionStatus.*;
 
 /**
  * Board의 CUD 비즈니스 로직을 담당하는 서비스 구현체
@@ -104,14 +103,13 @@ public class BoardServiceImpl implements BoardService {
 	@Override
 	public void deleteBoard(Long memberId, Long boardId) {
 		Board board = boardRepository.findById(boardId)
-				.orElseThrow(() -> new ServiceException(NOT_FOUND_BOARD));
+				.orElseThrow(NOT_FOUND_BOARD::asServiceException);
 		if (board.getDeletedAt() != null) {
-			throw new ServiceException(NOT_FOUND_BOARD);
+			throw ALREADY_DELETED_BOARD.asServiceException();
 		}
-		//TODO: 찾아온 board에서 boradId != memberId이면 ServiceException ?? -> 로직 확인 필요
-//		if (!board.isId(memberId)) {
-//			throw new ServiceException(UNAUTHENTICATED);
-//		}
+		if (!board.isOwnedBy(memberId)) {
+			throw UNAUTHENTICATED.asServiceException();
+		}
 		if (!board.getComments().isEmpty()) {
 			commentRepository.deleteAll(board.getComments());
 		}
